# api/api_server/swagger.yaml
definitions:
...
  api.ApiCartArticle:
    properties:
      count:
        type: integer
      product_id:
        type: integer
    type: object
...
  api.ApiPreviewCart:
    properties:
      price:
        $ref: '#/definitions/usecase.TotalPrice'
      products:
        items:
          $ref: '#/definitions/usecase.PreviewCartArticle'
        type: array
    type: object
...
  api.ApiProductIdentifier:
    properties:
      product_id:
        type: integer
    type: object
...
  api.ApiUpdateOrder:
    properties:
      order_id:
        type: integer
      status:
        type: string
    type: object
...
  errors.Error:
    properties:
      error:
        type: string
    type: object
  usecase.CartProductPrice:
    properties:
      base_cost:
        type: integer
      discount:
        type: integer
      total_cost:
        type: integer
    type: object
...
  usecase.PreviewCartArticle:
    properties:
      count:
        type: integer
      id:
        type: integer
      preview_image:
        type: string
      price:
        $ref: '#/definitions/usecase.CartProductPrice'
      title:
        type: string
    type: object
...
  usecase.TotalPrice:
    properties:
      total_base_cost:
        type: integer
      total_cost:
        type: integer
      total_discount:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /admin/order/status:
    post:
      consumes:
      - application/json
      description: Изменение статуса пользовательского заказа.
      parameters:
      - description: Данные заказа.
        in: body
        name: UpdateOrder
        required: true
        schema:
          $ref: '#/definitions/api.ApiUpdateOrder'
      produces:
      - application/json
      responses:
        "200":
          description: Статус успешно изменён.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Изменение статуса заказ.
  /cart:
    delete:
      consumes:
      - application/json
      description: Удаление пользовательской корзины.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Товары из пользовательской корзины успешно удалены.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Удаление корзины.
    get:
      consumes:
      - application/json
      description: Получение пользовательской корзины.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Товары из пользовательской корзины успешно получены.
          schema:
            $ref: '#/definitions/api.ApiPreviewCart'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Получение корзины.
  /cart/product:
    delete:
      consumes:
      - application/json
      description: Удаление товара из корзины юзера.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные товара.
        in: body
        name: ProductIdentifier
        required: true
        schema:
          $ref: '#/definitions/api.ApiProductIdentifier'
      produces:
      - application/json
      responses:
        "200":
          description: Товар успешно удалён.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Удаление товара из корзины.
    post:
      consumes:
      - application/json
      description: Добавление товара в корзину юзера.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные товара.
        in: body
        name: CartArticle
        required: true
        schema:
          $ref: '#/definitions/api.ApiCartArticle'
      produces:
      - application/json
      responses:
        "200":
          description: Товар успешно добавлен.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Добавление товара в корзину.
    put:
      consumes:
      - application/json
      description: Изменение продуктов в пользовательской корзине.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные товара.
        in: body
        name: CartArticle
        required: true
        schema:
          $ref: '#/definitions/api.ApiCartArticle'
      produces:
      - application/json
      responses:
        "200":
          description: Товар успешно удалён.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Изменение продуктов в корзине.
  /category:
    get:
      consumes:
      - application/json
      description: Получение каталога категорий.
      produces:
      - application/json
      responses:
        "200":
          description: Каталог категорий успешно получен.
          schema:
            items:
              $ref: '#/definitions/errors.Error'
            type: array
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение каталога категорий.
  /category/{id}:
    get:
      consumes:
      - application/json
      description: Получение подкатегории.
      parameters:
      - description: Id подкатегории
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Каталог категорий успешно получен.
          schema:
            items:
              $ref: '#/definitions/errors.Error'
            type: array
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение подкатегории.
...
swagger: "2.0"